Starting to refactor player class

+++ Items/Item.cs
@@ -21,7 +21,7 @@ namespace text_survival.Items
         public double Weight { get; set; } // in kg
         public Action<Player> UseEffect { get; set; }
         public string Description { get; set; } = "";
-        public int Quality { get; set; } // percentage 0% being extremely poor quality, 100% being perfect quality
+        public double Quality { get; set; } // percentage 0% being extremely poor quality, 100% being perfect quality
         public bool IsFound { get; set; }
         public IClonable<Item>.CloneDelegate Clone { get; set; }
